refactor(sroa): split the variant-return pass into its phases
4171 lines held three passes that shared three helpers. The root now
reads as the phase order and each phase is its own module:

- `layout` — how a variant packs into a result vector, the single source
  of truth both phases build on.
- `access` — turning `ref.test` / `ref.cast` / `struct.get` on the old
  boxed temp into reads of the flat locals.
- `wrapper` — resolving a call-site value through the `Block` / `Seq`
  wrappers lowering and inlining leave around it.
- `return_temp` — Phase 0, normalising NIR's return-only temps.
- `widen` — a whole function's boxed variant result to multi-value.
- `slot_flatten` — one nested `ref W` result slot to `W`'s own layout.

A pure move: every one of the 107 items is byte-identical modulo
whitespace. The `pub(super)` markers the move required now spell out the
cross-phase surface, which was previously indistinguishable from the
file's private interior.
